# Conflicts:
#	src/query/catalog/src/catalog/interface.rs
#	src/query/ee/Cargo.toml
#	src/query/service/src/catalogs/default/database_catalog.rs
#	src/query/service/src/catalogs/default/mutable_catalog.rs
#	src/query/service/src/catalogs/default/session_catalog.rs
#	src/query/service/tests/it/storages/fuse/operations/commit.rs
#	src/query/storages/fuse/src/fuse_table.rs
#	src/query/storages/fuse/src/io/locations.rs
#	tests/sqllogictests/suites/base/06_show/06_0014_show_table_functions.test

+++ src/common/base/src/runtime/metrics/registry.rs
@@ -43,8 +43,8 @@ use crate::runtime::metrics::process_collector::ProcessCollector;
 use crate::runtime::metrics::sample::MetricSample;
 use crate::runtime::ThreadTracker;
 
-pub const MIN_HISTOGRAM_BOUND: f64 = i64::MIN as f64;
-pub const MAX_HISTOGRAM_BOUND: f64 = i64::MAX as f64;
+pub const MIN_HISTOGRAM_BOUND: f64 = f64::MIN;
+pub const MAX_HISTOGRAM_BOUND: f64 = f64::MAX;
 
 pub static GLOBAL_METRICS_REGISTRY: LazyLock<GlobalRegistry> =
     LazyLock::new(GlobalRegistry::create);
